ten coomit

+++ frontend/src/screens/signupScreen/SignupScreen.jsx
@@ -4,39 +4,43 @@ import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
 import { Helmet } from 'react-helmet-async';
 import axios from 'axios';
-import './SigninScreen.css';
+import './SignupScreen.css';
 import { useContext, useEffect, useState } from 'react';
 import { Store } from '../../Store';
 import { toast } from 'react-toastify';
 import { getError } from '../../utils';
-const SigninScreen = () => {
+const SignupScreen = () => {
   const navigate = useNavigate();
 
   const { search } = useLocation();
   const redirectInUrl = new URLSearchParams(search).get('redirect');
   const redirect = redirectInUrl ? redirectInUrl : '/';
 
+  const [first_name, setFirstName] = useState('');
+  const [last_name, setLastName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
 
   const { state, dispatch: ctxDispatch } = useContext(Store);
   const { userInfo } = state;
 
   const submitHandler = async (e) => {
     e.preventDefault();
+    if (password !== confirmPassword) {
+      toast.error('Passwords do not match');
+      return;
+    }
     try {
       const { data } = await axios.post(
-        'http://localhost:3001/costumers/loginCustomer',
-        {
-          email,
-          password,
-        }
+        'http://localhost:3001/costumers/createCustomer',
+        { first_name, last_name, email, password }
       );
       ctxDispatch({ type: 'USER_SIGNIN', payload: data });
       localStorage.setItem('userInfo', JSON.stringify(data));
       navigate(redirect || '/');
     } catch (err) {
-      toast.error(getError(err))
+      toast.error(getError(err));
     }
   };
 
@@ -51,8 +55,24 @@ const SigninScreen = () => {
       <Helmet>
         <title>Sing In</title>
       </Helmet>
-      <h1 className="my-3">Sing In</h1>
+      <h1 className="my-3">Sign Up</h1>
       <Form onSubmit={submitHandler}>
+        <Form.Group className="mb-3" controlId="firstName">
+          <Form.Label>First Name</Form.Label>
+          <Form.Control
+            type="text"
+            required
+            onChange={(e) => setFirstName(e.target.value)}
+          />
+        </Form.Group>
+        <Form.Group className="mb-3" controlId="lastName">
+          <Form.Label>Last Name</Form.Label>
+          <Form.Control
+            type="text"
+            required
+            onChange={(e) => setLastName(e.target.value)}
+          />
+        </Form.Group>
         <Form.Group className="mb-3" controlId="email">
           <Form.Label>Email</Form.Label>
           <Form.Control
@@ -69,16 +89,24 @@ const SigninScreen = () => {
             onChange={(e) => setPassword(e.target.value)}
           />
         </Form.Group>
+        <Form.Group className="mb-3" controlId="confirmePassword">
+          <Form.Label> Confirm Password</Form.Label>
+          <Form.Control
+            type="password"
+            required
+            onChange={(e) => setConfirmPassword(e.target.value)}
+          />
+        </Form.Group>
         <div className="mb-3">
-          <Button type="submit">Sign In</Button>
+          <Button type="submit">Sign Up</Button>
         </div>
         <div className="mb-3">
-          New customer?{' '}
-          <Link to={`/signup?redirect=${redirect}`}>Create your acount</Link>
+          Already have an account?{' '}
+          <Link to={`/signin?redirect=${redirect}`}>Sgin-In</Link>
         </div>
       </Form>
     </Container>
   );
 };
 
-export default SigninScreen;
+export default SignupScreen;
